+++ react/utils/metrics/metrics.ts
@@ -12,15 +12,40 @@ type StopImpersonateMetric = {
   description: 'Stop Impersonate User Action - UI'
 }
 
+export type ChangeTeamMetric = {
+  kind: 'change-team-ui-event'
+  description: 'User change team/organization - UI'
+}
+
 export type Metric = {
   name: 'b2b-suite-buyerorg-data'
   account: string
-} & (ImpersonateMetric | StopImpersonateMetric)
+} & (ImpersonateMetric | StopImpersonateMetric | ChangeTeamMetric)
 
 export const sendMetric = async (metric: Metric) => {
   try {
     await axios.post(ANALYTICS_URL, metric)
   } catch (error) {
     console.warn('Unable to log metrics', error)
+  }
+}
+
+export type SessionResponseParam = {
+  namespaces: {
+    account: {
+      accountName: {
+        value: string
+      }
+    }
+    profile: {
+      email: {
+        value: string
+      }
+    }
+    authentication: {
+      storeUserEmail: {
+        value: string
+      }
+    }
   }
 }