+++ CHANGELOG.md
@@ -11,6 +11,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 - Added metrics to Impersonate events
 - Added metrics to Stop Impersonating events
+- Added metrics to change team events
 
 ## [1.25.0] - 2023-07-24